Export single notes and note-to-note relations as metronome marks
Refactored `findMetronome` to detect metronome markings without an
equals sign, allowing single-note tempo markings to be correctly
exported as `<metronome>` elements instead of `<words>`.

Updated `wordsMetronome` to handle cases where only a single beat-unit
is present, and improved support for note-to-note relations in
compliance with MusicXML 4.0.

Added a new round-trip test case `testMetronomeRefactor.xml`.

Fixed a potential SEGV in `midiInstrument` during MusicXML export by
adding comprehensive null checks for instruments, scores, and channel
mappings. Added a `const` version of `MasterScore::midiMapping()` to
allow safe access during export.

Fixed a signedness comparison warning in `findMetronome`.

+++ src/importexport/musicxml/internal/export/exportmusicxml.cpp
@@ -4856,7 +4856,7 @@ static bool findMetronome(const std::list<TextFragment>& list,
     }
 
     if (metroPos > 0 && words.at(metroPos - 1) == '(') {
-        if (words.size() > metroPos + metroLen && words.at(metroPos + metroLen) == ')') {
+        if (words.size() > size_t(metroPos + metroLen) && words.at(metroPos + metroLen) == ')') {
             hasParen = true;
             metroPos -= 1;
             metroLen += 2;
